+++ internal/client/registered_model_service.go
@@ -10,6 +10,20 @@ type CreateRegisteredModelFromCustomModelRequest struct {
 	RegisteredModelID string `json:"registeredModelId,omitempty"`
 }
 
+type CreateRegisteredModelFromLeaderboardRequest struct {
+	ModelID                       string   `json:"modelId"`
+	RegisteredModelName           *string  `json:"registeredModelName,omitempty"`
+	RegisteredModelID             *string  `json:"registeredModelId,omitempty"`
+	PredictionThreshold           *float64 `json:"predictionThreshold,omitempty"`
+	ComputeAllTsIntervals         *bool    `json:"computeAllTsIntervals,omitempty"`
+	DistributionPredictionModelID *string  `json:"distributionPredictionModelId,omitempty"`
+}
+
+type Tag struct {
+	Name  string `json:"name"`
+	Value string `json:"value"`
+}
+
 type RegisteredModel struct {
 	ID             string `json:"id"`
 	Name           string `json:"name"`
@@ -19,7 +33,8 @@ type RegisteredModel struct {
 }
 
 type UpdateRegisteredModelVersionRequest struct {
-	Name string `json:"name,omitempty"`
+	Name                string   `json:"name,omitempty"`
+	PredictionThreshold *float64 `json:"predictionThreshold,omitempty"`
 }
 
 type ListRegisteredModelsResponse struct {
@@ -32,13 +47,21 @@ type UpdateRegisteredModelRequest struct {
 }
 
 type RegisteredModelVersion struct {
-	ID                     string `json:"id"` // Registered model version ID
-	Name                   string `json:"name"`
-	BuildStatus            string `json:"buildStatus"`
-	ModelID                string `json:"modelId"`
-	RegisteredModelID      string `json:"registeredModelId"`
-	RegisteredModelVersion int    `json:"registeredModelVersion"`
-	Stage                  string `json:"stage"`
+	ID                     string                       `json:"id"` // Registered model version ID
+	Name                   string                       `json:"name"`
+	BuildStatus            string                       `json:"buildStatus"`
+	ModelID                string                       `json:"modelId"`
+	RegisteredModelID      string                       `json:"registeredModelId"`
+	RegisteredModelVersion int                          `json:"registeredModelVersion"`
+	Stage                  string                       `json:"stage"`
+	Target                 RegisteredModelVersionTarget `json:"target"`
+	Tags                   []Tag                        `json:"tags"`
+}
+
+type RegisteredModelVersionTarget struct {
+	Name                string   `json:"name"`
+	Type                string   `json:"type"`
+	PredictionThreshold *float64 `json:"predictionThreshold,omitempty"`
 }
 
 type ListRegisteredModelVersionsResponse struct {